atc: structure: add stub for `prompt` step
The plan is to add a new step called `prompt` which allows a user to
provide manual parameters to a job. These parameters will appear in the
job as a variable. I plan to extend the UI to provide an interface for
users to supply these parameters, but the API and CLI will also support
it.

concourse/concourse#9053

Signed-off-by: Sam Weaver <sweaver@relaypro.com>